roomsync/fastapi/requirements.txt
# FastAPI framework and ASGI server
fastapi==0.115.11
uvicorn==0.34.0

# ORM and Database tools
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
alembic==1.14.1
psycopg2-binary==2.9.9

# Data validation and settings management
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1

# OAuth and JWT libraries
Authlib==1.5.1
python-jose==3.4.0
bcrypt==4.1.2

# Other utilities (if needed)
gunicorn==23.0.0
httpx==0.28.1
